Physics Fix Alternate Proposal
- Changed MinimalChildDistance to an int value for performance reasons (will make sense with another added feature), distance can still be controlled granularly by ChildDistanceConnectionsEffect, so this shouldn't be an issue.
- Implemented new FastTanh method, is consistently 2x faster across all values, now without those pesky values approaching 0!
- Moved RequiredDistance out of CalcTanh to be used as a vector math operation instead, this should save a little performance
- Added Log2 lookuptable, this is needed to help keep the connection distance scaling under control. Thanks to it only requiring an int, this lookup method is much faster than regular log methods. Might want to look into profiling if reversing the statement order for the lookup table improves performance? Shouldn't be a big issue tho.
- As mentioned above, IdealDistance is now calculated by taking the log2 of the number of connections, so nodes with lots of connections are still well spaced, but won't be sent on an interstellar mission to find the next habitable exoplanet.

- Finally, the physics fix!!!

Explanation for Physics:

The problem-
Going back to the old (and traumatizing) days of my vector math classes, I noticed that the old implementation appeared to have two key issues. Firstly, subtracting from the component of each vector will always tend them towards one direction. If the force vector is smaller than the RequiredDistance variable (which it typically always is), it would effectively be setting every value of the force vector to some negative number, which explained why every value was going on a road trip to the farlands of the negative quadrants, as the force would always push them in the direction. The second issue was that these pull vectors aren't actually calculating the required pull, they are calculating the vector that represents the line between the two nodes, and their magnitude.

The Fix-
In order to fix this, I got the magnitude of the PullVector, and subtracted the IdealDistance. This means that when the PullVector's magnitude is greater than the IdealDistance (in other words, the ideal magnitude), the force vector will move the node to reduce the pull vectors magnitude, and vice versa when the IdealDistance is greater. The code actually does this in the opposite direction due to the pullvector having transform.position get subtracted instead of connection.*, but the transformation vector is the same, just in the other direction and offset into an imaginary part of the plane, but that stuff is confusing and not really important to know so I'm just going to not think about it :)

Performance seems good, behavior appears to work how I'd expect it should, please let me know if you have any other questions!